refactor: move rfid internals into sub-package
