+++ src/main/java/com/pi4j/crowpi/components/exceptions/RfidCollisionException.java
@@ -1,10 +1,10 @@
 package com.pi4j.crowpi.components.exceptions;
 
-import com.pi4j.crowpi.components.internal.MFRC522;
+import com.pi4j.crowpi.components.internal.rfid.PcdError;
 
 public class RfidCollisionException extends RfidException {
     public RfidCollisionException() {
-        super(MFRC522.PcdError.COLL_ERR);
+        super(PcdError.COLL_ERR);
     }
 
     public RfidCollisionException(String message) {
