+++ src/main/java/com/pi4j/crowpi/components/exceptions/RfidException.java
@@ -1,13 +1,13 @@
 package com.pi4j.crowpi.components.exceptions;
 
-import com.pi4j.crowpi.components.internal.MFRC522;
+import com.pi4j.crowpi.components.internal.rfid.PcdError;
 
 public class RfidException extends Exception {
     public RfidException(String message) {
         super(message);
     }
 
-    public RfidException(MFRC522.PcdError error) {
+    public RfidException(PcdError error) {
         super(error.getDescription());
     }
 }
