+++ src/main/java/com/pi4j/crowpi/components/internal/rfid/MFRC522.java
@@ -1,4 +1,4 @@
-package com.pi4j.crowpi.components.internal;
+package com.pi4j.crowpi.components.internal.rfid;
 
 import com.pi4j.crowpi.components.Component;
 import com.pi4j.crowpi.components.exceptions.RfidCollisionException;
@@ -103,12 +103,12 @@ public class MFRC522 extends Component {
         mifareStopCrypto1();
     }
 
-    protected Tag select() throws RfidException {
+    protected RfidUid select() throws RfidException {
         return select(0);
     }
 
     @SuppressWarnings("SameParameterValue")
-    private Tag select(int validBits) throws RfidException {
+    private RfidUid select(int validBits) throws RfidException {
         // Prepare buffer with 9 bytes length (7 byte UID + 2 byte CRC_A)
         // This buffer must contain the following structure:
         //      Byte 0: SEL         Cascade Level Selector
@@ -329,16 +329,16 @@ public class MFRC522 extends Component {
         }
 
         // Return new tag instance
-        return new Tag(uidBytes, uidSak);
+        return new RfidUid(uidBytes, uidSak);
     }
 
-    protected void mifareAuth(MifareKey key, byte blockAddr, Tag tag) throws RfidException {
+    protected void mifareAuth(MifareKey key, byte blockAddr, RfidUid uid) throws RfidException {
         // Prepare buffer for authentication command
         final byte[] buffer = new byte[12];
         buffer[0] = key.getType().getCommand().getValue();
         buffer[1] = blockAddr;
         System.arraycopy(key.getBytes(), 0, buffer, 2, Math.min(6, key.getLength()));
-        System.arraycopy(tag.getUid(), tag.getUidLength() - 4, buffer, 8, Math.min(4, tag.getUidLength()));
+        System.arraycopy(uid.getUid(), uid.getUidLength() - 4, buffer, 8, Math.min(4, uid.getUidLength()));
 
         System.out.println("Auth buffer: " + ByteHelpers.toString(buffer));
 
@@ -395,96 +395,6 @@ public class MFRC522 extends Component {
         }
         if (response.getBytes()[0] != PICC_MIFARE_ACK) {
             throw new RfidException("Received MIFARE NACK from PICC due to unknown error");
-        }
-    }
-
-    protected final static class MifareKey {
-        private final Type type;
-        private final byte[] bytes;
-
-        public MifareKey(Type type, byte[] bytes) {
-            if (bytes.length != 6) {
-                throw new IllegalArgumentException("Length of key must be exactly 6 bytes");
-            }
-
-            this.type = type;
-            this.bytes = bytes;
-        }
-
-        public static MifareKey getDefaultKeyA() {
-            return new MifareKey(Type.KEY_A, getDefaultKey());
-        }
-
-        public static MifareKey getDefaultKeyB() {
-            return new MifareKey(Type.KEY_B, getDefaultKey());
-        }
-
-        private static byte[] getDefaultKey() {
-            return new byte[]{(byte) 0xFF, (byte) 0xFF, (byte) 0xFF, (byte) 0xFF, (byte) 0xFF, (byte) 0xFF};
-        }
-
-        public byte[] getBytes() {
-            return bytes;
-        }
-
-        public int getLength() {
-            return bytes.length;
-        }
-
-        public Type getType() {
-            return type;
-        }
-
-        public enum Type {
-            KEY_A(PiccCommand.MF_AUTH_KEY_A),
-            KEY_B(PiccCommand.MF_AUTH_KEY_B);
-
-            private final PiccCommand command;
-
-            Type(PiccCommand command) {
-                this.command = command;
-            }
-
-            public PiccCommand getCommand() {
-                return command;
-            }
-        }
-    }
-
-    public final static class Tag {
-        private final byte[] uid;
-        private final byte sak;
-        private final String serial;
-
-        public Tag(byte[] uid, byte sak) {
-            this.uid = uid;
-            this.sak = sak;
-            this.serial = ByteHelpers.toString(uid);
-        }
-
-        public Tag(List<Byte> uidBytes, byte sak) {
-            this(ByteHelpers.toArray(uidBytes), sak);
-        }
-
-        public byte[] getUid() {
-            return uid;
-        }
-
-        public int getUidLength() {
-            return uid.length;
-        }
-
-        public byte getSak() {
-            return sak;
-        }
-
-        public String getSerial() {
-            return serial;
-        }
-
-        @Override
-        public String toString() {
-            return "Tag{uid=" + serial + ", sak=" + ByteHelpers.toString(sak) + "}";
         }
     }
 
@@ -752,538 +662,4 @@ public class MFRC522 extends Component {
         }
     }
 
-    /**
-     * Registers used for communicating with the PCD
-     */
-    @SuppressWarnings("SpellCheckingInspection")
-    private enum PcdRegister {
-        /**
-         * starts and stops command execution
-         */
-        COMMAND_REG(0x01),
-        /**
-         * enable and disable interrupt request control bits
-         */
-        COM_I_EN_REG(0x02),
-        /**
-         * enable and disable interrupt request control bits
-         */
-        DIV_I_EN_REG(0x02),
-        /**
-         * interrupt request bits
-         */
-        COM_IRQ_REG(0x04),
-        /**
-         * interrupt request bits
-         */
-        DIV_IRQ_REG(0x05),
-        /**
-         * error bits showing the error status of the last command executed
-         */
-        ERROR_REG(0x06),
-        /**
-         * communication status bits
-         */
-        STATUS_1_REG(0x07),
-        /**
-         * receiver and transmitter status bits
-         */
-        STATUS_2_REG(0x08),
-        /**
-         * input and output of 64 byte FIFO buffer
-         */
-        FIFO_DATA_REG(0x09),
-        /**
-         * number of bytes stored in the FIFO buffer
-         */
-        FIFO_LEVEL_REG(0x0A),
-        /**
-         * level for FIFO underflow and overflow warning
-         */
-        WATER_LEVEL_REG(0x0B),
-        /**
-         * miscellaneous control registers
-         */
-        CONTROL_REG(0x0C),
-        /**
-         * adjustments for bit-oriented frames
-         */
-        BIT_FRAMING_REG(0x0D),
-        /**
-         * bit position of the first bit-collision detected on the RF interface
-         */
-        COLL_REG(0x0E),
-
-        /**
-         * defines general modes for transmitting and receiving
-         */
-        MODE_REG(0x11),
-        /**
-         * defines transmission data rate and framing
-         */
-        TX_MODE_REG(0x12),
-        /**
-         * defines reception data rate and framing
-         */
-        RX_MODE_REG(0x13),
-        /**
-         * controls the logical behavior of the antenna driver pins TX1 and TX2
-         */
-        TX_CONTROL_REG(0x14),
-        /**
-         * controls the setting of the transmission modulation
-         */
-        TX_ASK_REG(0x15),
-        /**
-         * selects the internal sources for the antenna driver
-         */
-        TX_SEL_REG(0x16),
-        /**
-         * selects internal receiver settings
-         */
-        RX_SEL_REG(0x17),
-        /**
-         * selects thresholds for the bit decoder
-         */
-        RX_THRESHOLD_REG(0x18),
-        /**
-         * defines demodulator settings
-         */
-        DEMOD_REG(0x19),
-        /**
-         * controls some MIFARE communication transmit parameters
-         */
-        MF_TX_REG(0x1C),
-        /**
-         * controls some MIFARE communication receive parameters
-         */
-        MF_RX_REG(0x1D),
-        /**
-         * selects the speed of the serial UART interface
-         */
-        SERIAL_SPEED_REG(0x1F),
-
-        /**
-         * shows the MSB value of the CRC calculation
-         */
-        CRC_RESULT_REG_HIGH(0x21),
-        /**
-         * shows the LSB value of the CRC calculation
-         */
-        CRC_RESULT_REG_LOW(0x22),
-        /**
-         * controls the ModWidth setting
-         */
-        MOD_WIDTH_REG(0x24),
-        /**
-         * configures the receiver gain
-         */
-        RF_CFG_REG(0x26),
-        /**
-         * selects the conductance of the antenna driver pins TX1 and TX2 for modulation
-         */
-        GS_N_REG(0x27),
-        /**
-         * defines the conductance of the p-driver output during periods of no modulation
-         */
-        CW_GS_P_REG(0x28),
-        /**
-         * defines the conductance of the p-driver output during periods of modulation
-         */
-        MOD_GS_P_REG(0x29),
-        /**
-         * defines settings for the internal timer
-         */
-        T_MODE_REG(0x2A),
-        /**
-         * defines settings for the internal timer
-         */
-        T_PRESCALER_REG(0x2B),
-        /**
-         * defines the MSB of the 16-bit timer reload value
-         */
-        T_RELOAD_REG_HIGH(0x2C),
-        /**
-         * defines the LSB of the 16-bit timer reload value
-         */
-        T_RELOAD_REG_LOW(0x2D),
-        /**
-         * shows the MSB of the 16-bit timer value
-         */
-        T_COUNTER_VAL_REG_HIGH(0x2E),
-        /**
-         * shows the LSB of the 16-bit timer value
-         */
-        T_COUNTER_VAL_REG_LOW(0x2F);
-
-        private final byte value;
-        private final byte writeAddress;
-        private final byte readAddress;
-
-        PcdRegister(int value) {
-            this((byte) value);
-        }
-
-        PcdRegister(byte value) {
-            this.value = value;
-            this.writeAddress = (byte) ((value << 1) & 0x7E);
-            this.readAddress = (byte) (writeAddress | 0x80);
-        }
-
-        public byte getValue() {
-            return this.value;
-        }
-
-        public byte getReadAddress() {
-            return readAddress;
-        }
-
-        public byte getWriteAddress() {
-            return writeAddress;
-        }
-    }
-
-    public enum PcdError {
-        PROTOCOL_ERR(0x1, "Protocol error"),
-        PARITY_ERR(0x2, "Parity check failed"),
-        CRC_ERR(0x4, "CRC checksum mismatch"),
-        COLL_ERR(0x8, "Bit-collision detected"),
-        BUFFER_OVFL(0x10, "FIFO buffer overflow"),
-        TEMP_ERR(0x40, "Temperature error due to overheating"),
-        WR_ERR(0x80, "Illegal write error");
-
-        private final byte value;
-        private final String description;
-
-        PcdError(int value, String description) {
-            this((byte) value, description);
-        }
-
-        PcdError(byte value, String description) {
-            this.value = value;
-            this.description = description;
-        }
-
-        public static PcdError matchErrReg(byte errReg, PcdError... errors) {
-            for (final var error : errors) {
-                if (error.isSet(errReg)) {
-                    return error;
-                }
-            }
-            return null;
-        }
-
-        public byte getValue() {
-            return this.value;
-        }
-
-        public String getDescription() {
-            return this.description;
-        }
-
-        public boolean isSet(byte errReg) {
-            return (errReg & getValue()) != 0;
-        }
-
-        public boolean isClear(byte errReg) {
-            return !isSet(errReg);
-        }
-    }
-
-    /**
-     * COM IRQs used by the PCD
-     */
-    private enum PcdComIrq {
-        /**
-         * the timer decrements the timer value in register TCounterValReg to zero
-         */
-        TIMER_IRQ(0x1),
-        /**
-         * any error bit in the ErrorReg register is set
-         */
-        ERR_IRQ(0x2),
-        /**
-         * Status1Reg register’s LoAlert bit is set
-         */
-        LO_ALERT_IRQ(0x4),
-        /**
-         * the Status1Reg register’s HiAlert bit is set
-         */
-        HI_ALERT_IRQ(0x8),
-        /**
-         * if a command terminates, for example, when the CommandReg changes its value from any command to the Idle command.
-         * if an unknown command is started, the CommandReg register Command[3:0] value changes to the idle state and the IdleIRq bit is set.
-         * the microcontroller starting the Idle command does not set the IdleIRq bit.
-         */
-        IDLE_IRQ(0x10),
-        /**
-         * receiver has detected the end of a valid data stream
-         */
-        RX_IRQ(0x20),
-        /**
-         * set immediately after the last bit of the transmitted data was sent out
-         */
-        TX_IRQ(0x40);
-
-        private final byte value;
-
-        PcdComIrq(int value) {
-            this((byte) value);
-        }
-
-        PcdComIrq(byte value) {
-            this.value = value;
-        }
-
-        public byte getValue() {
-            return this.value;
-        }
-
-        public boolean isSet(byte comIrqReg) {
-            return (comIrqReg & getValue()) != 0;
-        }
-
-        public boolean isClear(byte comIrqReg) {
-            return !isSet(comIrqReg);
-        }
-    }
-
-    /**
-     * DIV IRQs used by the PCD
-     */
-    private enum PcdDivIrq {
-        /**
-         * the CalcCRC command is active and all data is processed
-         */
-        CRC_IRQ(0x4),
-        /**
-         * MFIN is active
-         * this interrupt is set when either a rising or falling signal edge is detected
-         */
-        MFIN_ACT_IRQ(0x10);
-
-        private final byte value;
-
-        PcdDivIrq(int value) {
-            this((byte) value);
-        }
-
-        PcdDivIrq(byte value) {
-            this.value = value;
-        }
-
-        public byte getValue() {
-            return this.value;
-        }
-
-        public boolean isSet(byte divIrqReg) {
-            return (divIrqReg & getValue()) != 0;
-        }
-
-        public boolean isClear(byte divIrqReg) {
-            return !isSet(divIrqReg);
-        }
-    }
-
-    /**
-     * Commands used for communicating with the PCD
-     */
-    @SuppressWarnings("SpellCheckingInspection")
-    private enum PcdCommand {
-        /**
-         * no action, cancels current command execution
-         */
-        IDLE(0b0000),
-        /**
-         * stores 25 bytes into the internal buffer
-         */
-        MEM(0b0001),
-        /**
-         * generates a 10-byte random ID number
-         */
-        GENERATE_RANDOM_ID(0b0010),
-        /**
-         * activates the CRC coprocessor or performs a self test
-         */
-        CALC_CRC(0b0011),
-        /**
-         * transmits data from the FIFO buffer
-         */
-        TRANSMIT(0b0100),
-        /**
-         * no command change, can be used to modify the CommandReg register bits without affecting the command, for example, the PowerDown bit
-         */
-        NO_CMD_CHANGE(0b0111),
-        /**
-         * activates the receiver circuits
-         */
-        RECEIVE(0b1000),
-        /**
-         * transmits data from FIFO buffer to antenna and automatically activates the receiver after transmission
-         */
-        TRANSCEIVE(0b1100),
-        /**
-         * performs the MIFARE standard authentication as a reader
-         */
-        MF_AUTHENT(0b1110),
-        /**
-         * resets the MFRC522
-         */
-        SOFT_RESET(0b1111);
-
-        private final byte value;
-
-        PcdCommand(int value) {
-            this((byte) value);
-        }
-
-        PcdCommand(byte value) {
-            this.value = value;
-        }
-
-        public byte getValue() {
-            return this.value;
-        }
-    }
-
-    /**
-     * Commands used for communicating with the PICC
-     */
-    @SuppressWarnings("SpellCheckingInspection")
-    private enum PiccCommand {
-        /**
-         * REQuest command, Type A.
-         * Invites PICCs in state IDLE to go to READY and prepare for anticollision or selection. 7 bit frame.
-         */
-        REQA(0x26),
-        /**
-         * HaLT command, Type A.
-         * Instructs an ACTIVE PICC to go to state HALT.
-         */
-        HLTA(0x50),
-        /**
-         * Wake-UP command, Type A.
-         * Invites PICCs in state IDLE and HALT to go to READY(*) and prepare for anticollision or selection. 7 bit frame.
-         */
-        WUPA(0x52),
-        /**
-         * Cascade Tag.
-         * Not really a command, but used during anti collision.
-         */
-        CASCADE_TAG(0x88),
-        /**
-         * Anti collision/Select, Cascade Level 1.
-         */
-        SEL_CL1(0x93),
-        /**
-         * Anti collision/Select, Cascade Level 2.
-         */
-        SEL_CL2(0x95),
-        /**
-         * Anti collision/Select, Cascade Level 3.
-         */
-        SEL_CL3(0x97),
-
-        /**
-         * [MIFARE Classic/Ultralight] Reads one 16 byte block from the authenticated sector of the PICC.
-         * Also used for MIFARE Ultralight.
-         */
-        MF_READ(0x30),
-        /**
-         * [MIFARE Classic/Ultralight] Writes one 16 byte block to the authenticated sector of the PICC.
-         * Called "COMPATIBILITY WRITE" for MIFARE Ultralight.
-         */
-        MF_WRITE(0xA0),
-        /**
-         * [MIFARE Classic] Perform authentication with Key A
-         */
-        MF_AUTH_KEY_A(0x60),
-        /**
-         * [MIFARE Classic] Perform authentication with Key B
-         */
-        MF_AUTH_KEY_B(0x61);
-
-        private final byte value;
-
-        PiccCommand(int value) {
-            this.value = (byte) value;
-        }
-
-        PiccCommand(byte value) {
-            this.value = value;
-        }
-
-        public byte getValue() {
-            return this.value;
-        }
-    }
-
-    private enum PiccCascadeLevel {
-        CL1(PiccCommand.SEL_CL1, 0, 4),
-        CL2(PiccCommand.SEL_CL2, 3, 7),
-        CL3(PiccCommand.SEL_CL3, 6);
-
-        private final PiccCommand command;
-        private final int uidIndex;
-        private final int nextThreshold;
-
-        PiccCascadeLevel(PiccCommand command, int uidIndex) {
-            this(command, uidIndex, 0);
-        }
-
-        PiccCascadeLevel(PiccCommand command, int uidIndex, int nextThreshold) {
-            this.command = command;
-            this.uidIndex = uidIndex;
-            this.nextThreshold = nextThreshold;
-        }
-
-        public PiccCascadeLevel getNext() {
-            return PiccCascadeLevel.values()[ordinal() + 1];
-        }
-
-        public PiccCommand getCommand() {
-            return command;
-        }
-
-        public int getUidOffset() {
-            return uidIndex;
-        }
-
-        public int getNextThreshold() {
-            return nextThreshold;
-        }
-    }
-
-    public static final class PiccResponse {
-        private final byte[] bytes;
-        private final int length;
-        private final int lastBits;
-
-        public PiccResponse(byte[] bytes, int length, int lastBits) {
-            this.bytes = bytes;
-            this.length = length;
-            this.lastBits = lastBits & 0x7;
-        }
-
-        public byte[] getBytes() {
-            return this.bytes;
-        }
-
-        public int getLength() {
-            return this.length;
-        }
-
-        public int getLastBits() {
-            return this.lastBits;
-        }
-
-        @Override
-        public String toString() {
-            return "PiccResponse{" +
-                "bytes=" + Arrays.toString(bytes) +
-                ", length=" + length +
-                ", lastBits=" + lastBits +
-                '}';
-        }
-    }
 }
