+++ src/main/java/com/pi4j/crowpi/components/internal/rfid/PcdDivIrq.java
@@ -1,0 +1,38 @@
+package com.pi4j.crowpi.components.internal.rfid;
+
+/**
+ * DIV IRQs used by the PCD
+ */
+enum PcdDivIrq {
+    /**
+     * the CalcCRC command is active and all data is processed
+     */
+    CRC_IRQ(0x4),
+    /**
+     * MFIN is active
+     * this interrupt is set when either a rising or falling signal edge is detected
+     */
+    MFIN_ACT_IRQ(0x10);
+
+    private final byte value;
+
+    PcdDivIrq(int value) {
+        this((byte) value);
+    }
+
+    PcdDivIrq(byte value) {
+        this.value = value;
+    }
+
+    public byte getValue() {
+        return this.value;
+    }
+
+    public boolean isSet(byte divIrqReg) {
+        return (divIrqReg & getValue()) != 0;
+    }
+
+    public boolean isClear(byte divIrqReg) {
+        return !isSet(divIrqReg);
+    }
+}
