+++ src/main/java/com/pi4j/crowpi/components/internal/rfid/PcdError.java
@@ -1,0 +1,48 @@
+package com.pi4j.crowpi.components.internal.rfid;
+
+public enum PcdError {
+    PROTOCOL_ERR(0x1, "Protocol error"),
+    PARITY_ERR(0x2, "Parity check failed"),
+    CRC_ERR(0x4, "CRC checksum mismatch"),
+    COLL_ERR(0x8, "Bit-collision detected"),
+    BUFFER_OVFL(0x10, "FIFO buffer overflow"),
+    TEMP_ERR(0x40, "Temperature error due to overheating"),
+    WR_ERR(0x80, "Illegal write error");
+
+    private final byte value;
+    private final String description;
+
+    PcdError(int value, String description) {
+        this((byte) value, description);
+    }
+
+    PcdError(byte value, String description) {
+        this.value = value;
+        this.description = description;
+    }
+
+    public static PcdError matchErrReg(byte errReg, PcdError... errors) {
+        for (final var error : errors) {
+            if (error.isSet(errReg)) {
+                return error;
+            }
+        }
+        return null;
+    }
+
+    public byte getValue() {
+        return this.value;
+    }
+
+    public String getDescription() {
+        return this.description;
+    }
+
+    public boolean isSet(byte errReg) {
+        return (errReg & getValue()) != 0;
+    }
+
+    public boolean isClear(byte errReg) {
+        return !isSet(errReg);
+    }
+}
